refactor: change screen file names
