+++ front/lib/app/config/routes.dart
@@ -3,8 +3,8 @@ import 'package:front/app/domain/presentation/home/home.dart';
 import 'package:front/app/domain/presentation/login/email_singin.dart';
 import 'package:front/app/domain/presentation/login/login.dart';
 import 'package:front/app/domain/presentation/login/signin.dart';
-import 'package:front/app/domain/presentation/project/screen/project_creation_page.dart';
-import 'package:front/app/domain/presentation/project/screen/project_detail_page.dart';
+import 'package:front/app/domain/presentation/project/screen/project_creation.dart';
+import 'package:front/app/domain/presentation/project/screen/project_detail.dart';
 import 'package:front/app/domain/presentation/team/screen/team_create.dart';
 import 'package:front/app/domain/presentation/team/screen/team_datail.dart';
 import 'package:front/app/domain/presentation/team/screen/teams_list.dart';
