+++ front/lib/app/domain/presentation/project/screen/project_detail.dart
